Add a helper to resolve test files paths
